% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ld_matrix.R
\name{calc_ld_matrix}
\alias{calc_ld_matrix}
\title{estimate LD matrix}
\usage{
calc_ld_matrix(G, POS2, info_snp, chrs = 1:22, ncores = 1)
}
\arguments{
\item{G}{genotypes from obj.bigSNP$genotypes}

\item{POS2}{vector of cM distances}

\item{info_snp}{result from runningn bigsnpr::snp_match between
a reference panel and the summary statistics.}

\item{chrs}{what chromosomes should be evaluated?}
}
\description{
estimate LD matrix
}
\note{
Chromosome 1 must always be included.
}
