% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR2.R
\name{construct_template}
\alias{construct_template}
\title{Construct templates for commandline clals}
\usage{
construct_template(program, default_args)
}
\arguments{
\item{program}{filepath of program to run}

\item{default_args}{a list of arguments, in the form of list("arg_name" = arg)}
}
\value{
a string
}
\description{
Construct templates for commandline clals
}
\examples{
\dontrun{
construct_template("/nfs/home/ldsc/ldsc.py", list("hm3" = "/home/arvhar/hapmap3.tsv"))
}
}
