% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{e_r2}
\alias{e_r2}
\title{Title Expected R2 in PRS.
Calculates the expected R2 in out of sample prediction.}
\usage{
e_r2(h, m = 50000, n)
}
\arguments{
\item{h}{SNP heritability of phenotype}

\item{m}{Number of independant SNPs used for predictor. Wray et al (2018) suggests
using m = 50000, and this is therefore the default parameter}

\item{n}{Sample size of discovery GWAS}
}
\value{
a float (0 -> 1)
}
\description{
Title Expected R2 in PRS.
Calculates the expected R2 in out of sample prediction.
}
\examples{
# the latest SCZ gwas has a effecive sample size of ~110k, and h2 of 0.24
# what do we expect the out of sample r2 to be?
e_r2(h = 0.24, m=50000, n = 110000)

}
