% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{file.exists.ext}
\alias{file.exists.ext}
\title{file exists with extension}
\usage{
file.exists.ext(fname, ext)
}
\arguments{
\item{fname}{string. file name}

\item{ext}{string. extension}
}
\value{
boolean
}
\description{
file exists with extension
}
