% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_templates.R
\name{gctb_sbayesr_default}
\alias{gctb_sbayesr_default}
\title{Title}
\usage{
gctb_sbayesr_default(
  infile,
  name,
  pi = "0.95,0.02,0.02,0.01",
  gamma = "0.0,0.01,0.1,1",
  chain_length = 10000,
  burn_in = 4000,
  thin = 10,
  out_freq = 10,
  seed = 2022,
  out = Sys.getenv("PRS_ARCHIVE"),
  ldmatrix = Sys.getenv("GCTB_LDMATRIX"),
  gctb = Sys.getenv("GCTB")
)
}
\arguments{
\item{infile}{discovery gwas in .ma format}

\item{name}{prefix for what the output file should be called. .snpRes will be added}

\item{pi}{pi paramter in gctb}

\item{gamma}{paramter in gctb}

\item{chain_length}{paramter in gctb}

\item{burn_in}{paramter in gctb}

\item{thin}{paramter in gctb}

\item{out_freq}{paramter in gctb}

\item{seed}{paramter in gctb}

\item{out}{paramter in gctb}

\item{ldmatrix}{location of ldmatrix filepaths, defaults by looking an environment}

\item{gctb}{location of gctb program}
}
\value{
bash code
}
\description{
Title
}
\examples{
\dontrun{
# if you have GCTB and GCTB_LDMATRIX  and prs_archivein your environemnt
gctb_sbayesr_default("filepath/to/my/gwas.ma", "my_name")
gctb_sbayesr_default("filepath/to/my/gwas.ma", "my_name",
out="/my/custom/output/directory")
}
}
