% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fit.R
\name{glm_prs_metrics}
\alias{glm_prs_metrics}
\title{Generate statistical metrics from a logistic regression using polygenic risk scores and case status}
\usage{
glm_prs_metrics(df, base_formula, formula, pop_prev, odds_ratio = FALSE)
}
\arguments{
\item{df}{a dataframe with prs, case status and covariates}

\item{base_formula}{a glm formula of the logistic regression with only covariates}

\item{formula}{a glm formula for the full regression (case status on prs and covariates)}

\item{pop_prev}{optional, for computing R2 on the liability scale.}

\item{odds_ratio}{should the estimates be convert to odds_ratio?}
}
\value{
a tibble
}
\description{
Generate statistical metrics from a logistic regression using polygenic risk scores and case status
}
\examples{
\dontrun{
# OBS: in the dataframe passed, outcome variable needs to be named case,
# and PRS column needs to be named score
base <- as.formula(case ~ PC1 + PC2 + sex)
full <- as.formula(case ~ score + PC1 + PC2 + sex)
glm_prs_metrics(df, base, full, pop_prev = 0.01)
}
}
