% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{liability_scale_h2}
\alias{liability_scale_h2}
\title{Title}
\usage{
liability_scale_h2(pop_prev, sample_prev, h2)
}
\arguments{
\item{pop_prev}{prevalence of the disorder in the general population}

\item{sample_prev}{the prevalence of the disorder in the sample}

\item{h2}{observed-scale heritability}
}
\value{
a float
}
\description{
Title
}
\examples{
liability_scale_h2(0.1, 0.5, 0.27)
}
