% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{liability_scale_r2}
\alias{liability_scale_r2}
\title{Convert a R2 estimate from observec scale to liability scale}
\usage{
liability_scale_r2(k, r2, p)
}
\arguments{
\item{k}{population disease risk}

\item{r2}{observed scale R2}

\item{p}{proportion of cases}
}
\value{
a float
}
\description{
Convert a R2 estimate from observec scale to liability scale
}
\examples{
liability_scale_r2(0.1, 0.15, 0.4)
}
