% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bigsnp_from_bed.R
\name{load_bigsnp_from_bed}
\alias{load_bigsnp_from_bed}
\title{load bigSNP from bed file}
\usage{
load_bigsnp_from_bed(bed, verbose = TRUE)
}
\arguments{
\item{bed}{file path to plink file WITH .bed extension}

\item{verbose}{print misc stats}
}
\description{
load bigSNP from bed file
}
