% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR.R
\name{make_bash_job}
\alias{make_bash_job}
\title{Make a bash job from inside R}
\usage{
make_bash_job(exe, ...)
}
\arguments{
\item{exe}{filepath for program to execute}

\item{...}{key-value pair arguments to apply to program}
}
\value{
a character vector
}
\description{
Make a bash job from inside R
}
\examples{
make_bash_job(exe = "/home/gcta64", bfile = "geno/bedbimfam", "maf" = 0.01, out = "/home/arvhar/results")
}
