% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR.R
\name{make_sbayesr_job}
\alias{make_sbayesr_job}
\title{Title}
\usage{
make_sbayesr_job(
  infile,
  pi = "0.95,0.02,0.02,0.01",
  gamma = "0.0,0.01,0.1,1",
  chain_length = 10000,
  burn_in = 4000,
  thin = 10,
  out_freq = 10,
  out,
  ldmatrix,
  gctb,
  name,
  from_env = TRUE
)
}
\arguments{
\item{infile}{filepath to a .ma file to be recaled}

\item{pi}{pi parameter for sbayesr}

\item{gamma}{gmama paramter for sbayesr}

\item{chain_length}{paramter for sbayers}

\item{burn_in}{paramter for sbayesr}

\item{thin}{parameter for sbayesr}

\item{out_freq}{parameters for sbayesr}

\item{out}{where to put output files}

\item{ldmatrix}{filepath for the ldmatrix needed for sbayesr}

\item{gctb}{filepath for sbayesr}

\item{name}{prefix name for all the output files}

\item{from_env}{if TRUE, the function looks up the filepaths by looking in
sys environment using GCTB, GCTB_LDMATRIX and PRS_ARCHIVE}
}
\value{
a string
}
\description{
Title
}
\examples{
\dontrun{
make_sbayesr_job("test_infile.ma", "my_out_dir", name="demo")
}
}
