% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR.R
\name{make_score_df}
\alias{make_score_df}
\title{Generate all combinations of the name column for two tibbles.}
\usage{
make_score_df(genotype, to_score)
}
\arguments{
\item{genotype}{a tibble with the column name, and filepaths for each genotype cohort}

\item{to_score}{a tibble with column name, and filepaths}
}
\value{
a tibble
}
\description{
Generate all combinations of the name column for two tibbles.
}
\examples{
\dontrun{

make_score_df(genotypes, to_score)
}
}
