% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR.R
\name{make_slurm_header}
\alias{make_slurm_header}
\title{Construct the slurm header for a script to be sent via sbatch}
\usage{
make_slurm_header(
  t = 480,
  cores = 4,
  mem = 10000,
  out = Sys.getenv("HOME"),
  name = "slurm"
)
}
\arguments{
\item{t}{TIME, d:h:m:s 00:16:30:30}

\item{cores}{number of cores to be used}

\item{mem}{memory in megabytes}

\item{out}{where to save the slurm output file}

\item{name}{the name of the script}
}
\value{
a string
}
\description{
Construct the slurm header for a script to be sent via sbatch
}
\examples{
make_slurm_header()
}
