% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{neff}
\alias{neff}
\title{Title Calculate the effective sample size}
\usage{
neff(cases, controls)
}
\arguments{
\item{cases}{number of affected individuals}

\item{controls}{number of controls}
}
\value{
a float
}
\description{
Title Calculate the effective sample size
}
\examples{
n_eff <- neff(25000, 75000)
}
