% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_templates.R
\name{plink_score_default}
\alias{plink_score_default}
\title{Title}
\usage{
plink_score_default(bfile, score, out, threads = 16, format = "pgen")
}
\arguments{
\item{bfile}{filepath to genotype data}

\item{score}{filepath to sumstat, a1 is column 2, column 5 is effect size}

\item{out}{output directory and filename}

\item{threads}{number of threads}

\item{format}{bgen/pgen/etc}
}
\value{
bash code
}
\description{
Title
}
\examples{
\dontrun{
plink_score_default("/cohort_x/name_genotypes", "my/rescaled/gwas.ma", "/store/prs/here/name")
}
}
