% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hail_sumstat.R
\name{read_hail_sumstat}
\alias{read_hail_sumstat}
\title{read hail summary statistics}
\usage{
read_hail_sumstat(path)
}
\arguments{
\item{path}{string file to summary statistics generated with HAIL}
}
\value{
data.table of ldpred2 ready summary statistics
}
\description{
read hail summary statistics
}
