% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_gwas_catalogue.R
\name{select_cols}
\alias{select_cols}
\title{Select columns from gwas catalogue, to convert to .ma format from GCTA}
\usage{
select_cols(df)
}
\arguments{
\item{df}{dataframe downloaded from GWAS catalogue}
}
\value{
a tibble
}
\description{
Select columns from gwas catalogue, to convert to .ma format from GCTA
}
\examples{
\dontrun{
select_cols(df)
}
}
