% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_as_genetic_position.R
\name{snp_as_genetic_position}
\alias{snp_as_genetic_position}
\title{SNP as genetic position}
\usage{
snp_as_genetic_position(
  infos.chr,
  infos.pos,
  mapdir,
  ncores = 1,
  rsid = NULL,
  genetic_map = "omni"
)
}
\arguments{
\item{infos.chr}{vector of chromosomes of length n}

\item{infos.pos}{vector of genetic positions of length n}

\item{mapdir}{What directory should genetic maps (interpolated dist) be found at?}

\item{ncores}{how many cores should be used during computation}
}
\value{
vector of genetic positions of length n
}
\description{
SNP as genetic position
}
