% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR2.R
\name{to_bash}
\alias{to_bash}
\title{Title}
\usage{
to_bash(..., slurm = FALSE)
}
\arguments{
\item{...}{name-value pairs of arguments}

\item{slurm}{use a slurm prefix or not?}
}
\value{
a string
}
\description{
Title
}
\examples{
\dontrun{
to_bash("bfile" = "my_genetic/data")
}
}
