% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurmR.R
\name{make_slurm_header}
\alias{make_slurm_header}
\title{Title}
\usage{
make_slurm_header(...)
}
\arguments{
\item{...}{key-value pairs of slurm flags. Accepts both (ntasks = 6) and
shortform (-N 6)}
}
\value{
a character vector
}
\description{
Title
}
\examples{
make_slurm_header("-A" = "sens2017519", ntasks = 6, "-T" = "24:00:00")
}
